move lib files
Move some of the lib files into a separate directory to better organize things.
