+++ lib/game/Assignment.ts
