+++ lib/game/README.md
